Rename some files in the "database" folder.
